modules/mod-009-monitoring-basics/exercise-06-airflow-workflow-monitoring/requirements.txt
# Airflow and core dependencies
apache-airflow==2.7.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Data processing (for ML tasks)
pandas==2.1.3
numpy==1.24.3

# Machine learning
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
